requirements.txt
flask
pandas
pandasql
mysql-connector-python
google-generativeai
openpyxl